Package + test custom delimiter
