+++ src/Calculator/CalculatorTest.java
@@ -3,6 +3,8 @@ import static org.junit.Assert.*;
 import org.junit.Before;
 import org.junit.Test;
 
+package Calculator;
+
 public class CalculatorTest {
 	// Constants to pass to add()
 	private static final String EMPTY_STRING = "";
